// typechain-types/contracts/Vault.ts
export declare namespace Vault {
  export type LockedTokenStruct = {
    amount: PromiseOrValue<BigNumberish>;
    unlockTime: PromiseOrValue<BigNumberish>;
    released: PromiseOrValue<boolean>;
  };

  export type LockedTokenStructOutput = [BigNumber, BigNumber, boolean] & {
    amount: BigNumber;
    unlockTime: BigNumber;
    released: boolean;
  };
}